requirements.txt
pillow==11.1.0
matplotlib==3.10.1
numpy==2.2.4
pyyaml==6.0.2
scipy==1.15.2
scikit-learn==1.6.1
packaging==24.2
streamlit==1.44.0
torch==2.6.0
torchvision==0.21.0
pytorch_msssim==1.0.0
tqdm==4.67.1
opencv-python==4.11.0.86
imageio==2.37.0